rename `withNextRouter.` to `withTestRouter` to keep the same name as the function
