+++ stories/components/AppNav.stories.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { MockedProvider } from '@apollo/client/testing'
 import GET_APP from '../../graphql/queries/getApp'
 import AppNav from '../../components/AppNav'
-import { withTestRouter } from '../../__tests__/utils/withNextRouter'
+import { withTestRouter } from '../../__tests__/utils/withTestRouter'
 
 export default {
   component: AppNav,
