+++ stories/pages/ResetPassword.stories.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import ResetPassword from '../../pages/confirm/[token]'
 import { MockedProvider } from '@apollo/client/testing'
 import GET_APP from '../../graphql/queries/getApp'
-import { withTestRouter } from '../../__tests__/utils/withNextRouter'
+import { withTestRouter } from '../../__tests__/utils/withTestRouter'
 
 export default {
   component: ResetPassword,
